# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant RAG & Jev AI Service
  description: Multi-tenant knowledge indexing, hybrid retrieval and grounded generation. Tenant comes from the trusted context;
    Jev orchestrates capability only.
  version: 1.0.0
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      description: Tukar kode akses dengan sesi. Kode salah ditolak sebelum sesi apa pun dibuat.
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Login Api V1 Auth Login Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/session:
    get:
      tags:
      - auth
      summary: Read Session
      description: 'Siapa yang sedang masuk: kunci API, sesi konsol, atau token KMS.'
      operationId: read_session_api_v1_auth_session_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Read Session Api V1 Auth Session Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - auth
      summary: Logout
      description: 'Keluar: sesi yang sedang dipakai dicabut. Kunci API tidak bisa ''keluar'' di sini.'
      operationId: logout_api_v1_auth_session_delete
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Logout Api V1 Auth Session Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout Post
      description: Sama seperti ``DELETE /auth/session``; disediakan untuk klien yang hanya bisa POST.
      operationId: logout_post_api_v1_auth_logout_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Tenant-Context
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Tenant-Context
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Logout Post Api V1 Auth Logout Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        code:
          type: string
          maxLength: 256
          minLength: 1
          title: Code
        remember:
          type: boolean
          title: Remember
          default: false
      additionalProperties: false
      type: object
      required:
      - code
      title: LoginRequest
      description: 'Kode akses + pilihan "ingat saya". Tidak ada field tenant: konteks dibentuk server.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError